+++ server/video/src/main/java/com/github/makewheels/video2022/watch/play/WatchLog.java
@@ -17,6 +17,7 @@ public class WatchLog {
     private String videoId;
     private String clientId;
     private String sessionId;
+    private String viewerId;
     private String userAgent;
     private String videoStatus;