+++ server/video/src/main/java/com/github/makewheels/video2022/watch/play/WatchService.java
@@ -78,6 +78,7 @@ public class WatchService {
         watchLog.setVideoId(video.getId());
         watchLog.setClientId(context.getClientId());
         watchLog.setSessionId(context.getSessionId());
+        watchLog.setViewerId(UserHolder.getUserId());
 
         mongoTemplate.save(watchLog);
     }